requirements.txt
feedparser>=6.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
transformers>=4.36.0
streamlit>=1.29.0
torch>=2.1.0
python-dotenv>=1.0.0
scheduler>=0.8.0